refactor(var): restructure the module
