--- packages/utilities/var/src/construction/construction-detection-hint.ts
@@ -1,6 +1,0 @@
-export type ConstructionDetectionHint = bigint;
-
-export const ConstructionDetectionHint = {
-  None: 0n,
-  StrictCallExpression: 1n << 0n,
-};
