+++ packages/utilities/var/src/index.ts
@@ -1,9 +1,7 @@
-export * from "./construction";
-export * from "./find-property-in-properties";
-export * from "./find-variable";
-export * from "./get-child-scopes";
-export * from "./get-variable-declarator-id";
-export * from "./get-variable-init-node";
-export * from "./get-variables";
-export * from "./is-node-value-equal";
-export * from "./lazy-value";
+export * from "./var-collect";
+export * from "./var-construction";
+export * from "./var-declarator-id";
+export * from "./var-init-node";
+export * from "./var-scope";
+export * from "./var-value";
+export * from "./var-value-equal";
