+++ packages/utilities/var/src/var-scope.ts
